# docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GW2 WvW日志系统",
    "description": "GW2 WvW日志解析与评分系统后端API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "tags": [
          "日志管理"
        ],
        "summary": "获取日志列表",
        "operationId": "get_logs_api_v1_logs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "parse_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "解析状态",
              "title": "Parse Status"
            },
            "description": "解析状态"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "日志管理"
        ],
        "summary": "上传日志文件",
        "operationId": "upload_log_api_v1_logs_post",
        "parameters": [
          {
            "name": "auto_parse",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否自动解析",
              "default": false,
              "title": "Auto Parse"
            },
            "description": "是否自动解析"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_log_api_v1_logs_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wvw-report/logs": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取有同步数据的日志列表",
        "description": "获取已有 ZEVTC→EI 同步数据的日志列表",
        "operationId": "list_wvw_reports_api_v1_wvw_report_logs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/summary": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取战斗概览",
        "description": "返回战斗概览，包含：\n  - 基础信息（时长、人数、目标数）\n  - 伤害统计（总伤害、直伤、症状、平均DPS）\n  - 防御统计（倒地、死亡、承伤）\n  - 支援统计（症状清除、治疗）\n  - 小队构成（职业分布、小队分配）\n  - 阶段列表",
        "operationId": "get_wvw_summary_api_v1_wvw_report_logs__log_id__summary_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/players": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取玩家排行榜",
        "description": "返回所有玩家的摘要数据，支持多种排序方式：\n  - damage: 按总伤害排序\n  - dps: 按 DPS 排序\n  - taken: 按承受伤害排序\n  - downs: 按倒地次数排序\n  - support: 按症状清除排序\n  - healing: 按治疗量排序",
        "operationId": "get_wvw_players_api_v1_wvw_report_logs__log_id__players_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序字段: damage/dps/taken/downs/support/healing",
              "default": "damage",
              "title": "Sort By"
            },
            "description": "排序字段: damage/dps/taken/downs/support/healing"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwPlayersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/players/{player_id}": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取单个玩家详情",
        "description": "返回单个玩家的完整数据（含 DPS、防御、支援、技能循环等）",
        "operationId": "get_wvw_player_detail_api_v1_wvw_report_logs__log_id__players__player_id__get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          },
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "玩家ID（ei_player.player_id）",
              "title": "Player Id"
            },
            "description": "玩家ID（ei_player.player_id）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwPlayerDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/targets": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取目标列表",
        "description": "返回所有目标（NPC/敌方玩家）的摘要数据",
        "operationId": "get_wvw_targets_api_v1_wvw_report_logs__log_id__targets_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwTargetsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/phases": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取阶段列表",
        "description": "返回所有阶段的统计数据",
        "operationId": "get_wvw_phases_api_v1_wvw_report_logs__log_id__phases_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwPhasesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/timeline": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取战斗时间线",
        "description": "返回战斗关键事件时间线：\n  - 进入/退出战斗\n  - 倒地/起身\n  - Despawn 事件",
        "operationId": "get_wvw_timeline_api_v1_wvw_report_logs__log_id__timeline_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wvw-report/logs/{log_id}/skill-map": {
      "get": {
        "tags": [
          "WvW 战斗报告"
        ],
        "summary": "获取技能映射表",
        "description": "返回技能 ID 到名称的映射表",
        "operationId": "get_wvw_skill_map_api_v1_wvw_report_logs__log_id__skill_map_get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "日志ID",
              "title": "Log Id"
            },
            "description": "日志ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WvwSkillMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示消息"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "错误代码"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "时间戳"
          },
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "HTTP状态码"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ApiResponse"
      },
...
      "Body_upload_log_api_v1_logs_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_log_api_v1_logs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WvwPhasesResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwPhasesResponse"
      },
      "WvwPlayerDetailResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwPlayerDetailResponse"
      },
      "WvwPlayersResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwPlayersResponse"
      },
      "WvwReportListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwReportListResponse"
      },
      "WvwSkillMapResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwSkillMapResponse"
      },
      "WvwSummaryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwSummaryResponse"
      },
      "WvwTargetsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwTargetsResponse"
      },
      "WvwTimelineResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "WvwTimelineResponse"
      }
    }
  }
}